development/projects/GPT-Agents/autogen_cache_practice/requirements.txt
pyautogen[openai]>=0.2.0
python-dotenv>=0.19.0
openai>=1.0.0 